Added path correction

+++ libraries/roverLib/Ultrasonic.cpp
@@ -19,6 +19,6 @@ float Ultrasonic::read(){
     delayMicroseconds(10);
     digitalWrite(trigPin, LOW);
     long duration = pulseIn(echoPin, HIGH);
-    int distance = duration * 0.034 / 2;
+    float distance = duration * 0.034 / 2;
     return distance;
 }
